# docs/openapi.yaml
openapi: 3.1.0
info:
  title: muse-replicate Managed Agents API
  version: 0.1.0
  description: |
    企业私有化 Managed Agents 平台（Meta Muse 复刻）的 HTTP API，方言为 BigModel Managed Agents 子集。

    **鉴权**：`Authorization: Bearer mas_sk_...` 或 `x-api-key: mas_sk_...`（任选其一）。

    **方言 headers**（建议随请求发送）：
    - `zai-version: 2026-05-26`
    - `zai-beta: managed-agents-2026-05-26`

    兼容 Anthropic 方言（`anthropic-version` / `anthropic-beta` / `x-api-key`），服务端按 header 自动检测方言并适配错误信封。

    **错误信封**（所有非 2xx）：
    ```json
    { "type": "error", "error": { "type": "invalid_request_error", "message": "...", "details": {} }, "request_id": "req_..." }
    ```

    **通用语义**：
    - 所有 POST 接受 `Idempotency-Key` header（同 key 同 body 回放首次响应，异 body 409 `idempotency_conflict`）；
    - 列表接口支持 `limit` / 分页游标（`next_page`），Files 用 `before_id`/`after_id`；
    - 事件日志是唯一事实来源：会话事件按 `seq` 单调定序；SSE 见 `GET /v1/sessions/{id}/events/stream`。
servers:
  - url: http://127.0.0.1:8080
    description: 本地部署（PORT 环境变量覆盖）
...
security:
  - bearerAuth: []
  - apiKeyHeader: []
paths:
...
  /v1/sessions/{id}/events:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    post:
      tags: [Events]
      operationId: sendEvents
      summary: 发送用户事件（1-10 条；排队 processed_at=null；interrupt/confirmation 状态前置校验）
      parameters: *dialect
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventsSend" }
      responses:
        "200":
          description: 准入结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/SessionEvent" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      tags: [Events]
      operationId: listEvents
      summary: 历史事件（types/order/created_at/after_seq/limit≤100；排队事件排在 asc 末尾）
      parameters:
        - $ref: "#/components/parameters/ZaiVersion"
        - $ref: "#/components/parameters/ZaiBeta"
        - { name: types, in: query, schema: { type: string }, description: "逗号分隔事件类型" }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
        - { name: after_seq, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 100 } }
        - { name: "created_at[gt]", in: query, schema: { type: string, format: date-time } }
        - { name: "created_at[gte]", in: query, schema: { type: string, format: date-time } }
        - { name: "created_at[lt]", in: query, schema: { type: string, format: date-time } }
        - { name: "created_at[lte]", in: query, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: 事件列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/SessionEvent" } }
                  next_page: { type: [string, "null"] }
  /v1/sessions/{id}/events/stream:
    get:
      tags: [Events]
      operationId: streamEvents
      summary: SSE 实时流（LISTEN/NOTIFY、15s 心跳、Last-Event-ID 回补、session.deleted 关闭）
      parameters:
        - $ref: "#/components/parameters/ZaiVersion"
        - $ref: "#/components/parameters/ZaiBeta"
        - { name: Last-Event-ID, in: header, schema: { type: string }, description: "从该事件 id 起回补" }
      responses:
        "200":
          description: "text/event-stream（event: 事件类型；data: 事件 JSON；: ping 心跳）"
          content:
            text/event-stream:
              schema: { type: string }
  /v1/files:
    post:
      tags: [Files]
      operationId: uploadFile
      summary: multipart 上传文件（FILE-01；sha256 服务端计算；文件名黑名单 FILE-02）
      parameters: *dialect
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: 上传成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileObject" }
        "400": { $ref: "#/components/responses/BadRequest" }
    get:
      tags: [Files]
      operationId: listFiles
      summary: 文件列表（before_id/after_id 分页；scope 过滤）
      parameters:
        - $ref: "#/components/parameters/ZaiVersion"
        - $ref: "#/components/parameters/ZaiBeta"
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: before_id, in: query, schema: { type: string } }
        - { name: after_id, in: query, schema: { type: string } }
        - { name: session_id, in: query, schema: { type: string } }
      responses:
        "200":
          description: 文件列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/FileObject" } }
                  next_page: { type: [string, "null"] }
  /v1/files/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [Files]
      operationId: getFile
      summary: 文件元数据
      parameters: *dialectGet
      responses:
        "200":
          description: 文件元数据
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileObject" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Files]
      operationId: deleteFile
      summary: 删除文件
      parameters: *dialectGet
      responses:
        "200": { description: 删除成功 }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/files/{id}/content:
    get:
      tags: [Files]
      operationId: downloadFileContent
      summary: 下载文件内容（Content-Type 透传）
      parameters: *dialectGet
      responses:
        "200":
          description: 文件字节流
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Authorization: Bearer mas_sk_..."
    apiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
      description: "x-api-key: mas_sk_...（BigModel/Anthropic 方言兼容）"
  parameters:
    ZaiVersion:
      name: zai-version
      in: header
      required: false
      schema: { type: string, default: "2026-05-26" }
      description: BigModel API 版本（等价 anthropic-version）
    ZaiBeta:
      name: zai-beta
      in: header
      required: false
      schema: { type: string, default: "managed-agents-2026-05-26" }
      description: Managed Agents beta 通道（等价 anthropic-beta）
...
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: 请求体/参数不合法（invalid_request_error）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    NotFound:
      description: 资源不存在（not_found_error）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: 状态冲突 / 乐观锁 / 幂等冲突 / 引用约束（conflict_error）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    Error:
      type: object
      required: [type, message]
      properties:
        type:
          type: string
          enum: [invalid_request_error, authentication_error, permission_error, not_found_error, conflict_error, rate_limit_error, api_error, overloaded_error]
        message: { type: string }
        details:
          type: [object, "null"]
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [type, error, request_id]
      properties:
        type: { const: error }
        error: { $ref: "#/components/schemas/Error" }
        request_id: { type: string }
    Timestamp:
      type: string
      format: date-time
...
    SessionEvent:
      type: object
      description: payload 在顶层展开（content/subtype/usage 等）；排队事件 seq/processed_at 为 null
      required: [id, type]
      properties:
        id: { type: string }
        type: { type: string, examples: [user.message, agent.message, agent.custom_tool_use, user.custom_tool_result, session.status_running, session.status_idle, session.usage] }
        seq: { type: [integer, "null"] }
        processed_at: { type: [string, "null"], format: date-time }
        received_at: { $ref: "#/components/schemas/Timestamp" }
    EventsSend:
      type: object
      required: [events]
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 10
          description: 整批原子；仅用户事件；requires_action 时只接受 tool_confirmation/interrupt
          items:
            oneOf:
              - type: object
                required: [type, content]
                properties:
                  type: { const: user.message }
                  content:
                    type: array
                    minItems: 1
                    maxItems: 20
                    items:
                      type: object
                      additionalProperties: true
                      required: [type]
                      properties:
                        type: { type: string, enum: [text, image, document] }
                        text: { type: string }
              - type: object
                required: [type]
                properties:
                  type: { const: user.interrupt }
              - type: object
                required: [type, tool_use_id, output]
                description: 自定义工具结果（spec §7.3 例外：收到即处理，返回时 processed_at 已有值）
                properties:
                  type: { const: user.custom_tool_result }
                  tool_use_id: { type: string, description: 对应 agent.custom_tool_use 事件 id（须在 requires_action 的 event_ids 中） }
                  output: { type: string, minLength: 1, maxLength: 64000 }
              - type: object
                required: [type, tool_use_id, result]
                properties:
                  type: { const: user.tool_confirmation }
                  tool_use_id: { type: string }
                  result: { type: string, enum: [allow, deny] }
                  deny_message: { type: string, maxLength: 4000 }
    FileObject:
      type: object
      required: [id, type, filename, bytes]
      properties:
        id: { type: string }
        type: { const: file }
        filename: { type: string }
        bytes: { type: integer }
        content_type: { type: [string, "null"] }
        sha256: { type: string }
        scope: { type: [string, "null"] }
        created_at: { $ref: "#/components/schemas/Timestamp" }
...